mqtt_handler/requirements.txt
paho-mqtt
pymongo
mysql-connector-python
neo4j